requirements.txt
opencv-python
numpy
pillow
streamlit
ultralytics
torch
torchvision
torchaudio
pyyaml
matplotlib
seaborn
